requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.6,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.26,<1.0
Pillow>=10.0,<12.0
pydub>=0.25,<1.0
pytest>=8.0,<9.0

